+++ projects/deario/handlers/push.go
@@ -1,0 +1,39 @@
+package handlers
+
+import (
+	"net/http"
+
+	"simple-server/pkg/util/authutil"
+	"simple-server/projects/deario/db"
+
+	"github.com/labstack/echo/v4"
+)
+
+// SavePushKey는 푸시 토큰을 저장한다.
+func SavePushKey(c echo.Context) error {
+	uid, err := authutil.SessionUID(c)
+	if err != nil {
+		return err
+	}
+
+	var data map[string]interface{}
+	if err := c.Bind(&data); err != nil {
+		return err
+	}
+
+	token := data["token"].(string)
+
+	queries, err := db.GetQueries(c.Request().Context())
+	if err != nil {
+		return err
+	}
+
+	if err := queries.UpsertPushKey(c.Request().Context(), db.UpsertPushKeyParams{
+		Uid:       uid,
+		PushToken: token,
+	}); err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, "푸시 키 저장 실패")
+	}
+
+	return nil
+}
